shared / back / front  - move createArchivedConventionRequest route from magicLink routes to authenticated routes
