+++ back/src/adapters/primary/routers/magicLink/createMagicLinkRouter.ts
@@ -61,17 +61,6 @@ export const createMagicLinkRouter = (deps: AppDependencies) => {
       ),
   );
 
-  sharedRouter.createArchivedConventionRequest(
-    deps.connectedUserAuthMiddleware,
-    (req, res) =>
-      sendHttpResponse(req, res.status(201), () =>
-        deps.useCases.createArchivedConventionRequest.execute(
-          req.body,
-          getGenericAuthOrThrow(req.payloads?.currentUser),
-        ),
-      ),
-  );
-
   sharedRouter.updateConvention(
     deps.conventionMagicLinkAuthMiddleware,
     (req, res) =>
